Rename: 컴포넌트 재명명
